refactor: move partner integrations components to pages components folder
