+++ apps/web/src/pages/partner-integrations/components/ProjectRow.tsx
